♻️  refactor: 폴더 구조 및 변수명 수정
