+++ app/layout.tsx
@@ -2,10 +2,7 @@ import type { Metadata } from "next";
 import "@/styles/globals.css";
 
 export const metadata: Metadata = {
-  title: {
-    template: "%s | 급Pay",
-    default: "급Pay",
-  },
+  title: "급Pay",
   description: "급하게 일손이 필요한 자리에 더 많은 시급을 제공해서 아르바이트생을 구할 수 있는 서비스",
 };
