+++ constants/user-type.ts
@@ -1,6 +1,6 @@
-export const userType = {
+export const USER_TYPE = {
   EMPLOYEE: "employee",
   EMPLOYER: "employer",
 } as const;
 
-export type UserType = (typeof userType)[keyof typeof userType];
+export type UserType = (typeof USER_TYPE)[keyof typeof USER_TYPE];
